moving test_one_sample
